moved MimeMessageBuilder back from test to main. Adjusted README.
